requirements/prod.txt
# Everything needed in production

# Flask
Flask==0.12.2
MarkupSafe==1.0
Werkzeug==0.12.2
Jinja2==2.9.6
itsdangerous==0.24
click>=5.0

# Database
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
SQLAlchemy==1.1.14
records>=0.5.2

# Migrations
Flask-Migrate==2.1.1

# Forms
Flask-WTF==0.14.2
WTForms==2.1

# Deployment
gunicorn>=19.1.1

# Webpack
flask-webpack==0.1.0

# Auth
Flask-Login==0.4.0
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar==0.10.1

# HTTP
requests>=2.9.1

# Cryptocurrency
cbpro==1.1.2

# Configs
pyyaml>=3.12

# Data
pandas>=0.21
pandas-datareader>=0.1
